create middleware to add custom response headers. add server api version header. rename exceptions.py errors.py and move error creation functions there.
