+++ robot-server/robot_server/__init__.py
@@ -1,0 +1,3 @@
+# This is the version of the API supported by the server. The
+# application version is in package.json.
+API_VERSION = "1.0.0"
